+++ echo-OJT-webApp/app/Dockerfile
@@ -10,7 +10,7 @@ LABEL maintainer="Kataru Onosato <konosato@idcf.jp>"
 WORKDIR /app
 
 # Copy go mod and sum files
-COPY go.mod go.sum ./
+COPY ../go.mod go.sum ./
 
 # Download all dependencies. Dependencies will be cached if the go.mod and go.sum files are not changed
 RUN go mod download
